+++ src/oa_cohorts/measurables/ev_measureables.py
@@ -1,4 +1,4 @@
-from omop_constructs.alchemy.events import DxMeasurementMV, DxProcedureMV, DxObservationMV
+from omop_constructs.alchemy.events import DxMeasurementMV, DxProcedureMV, DxObservationMV, DxRelevantVisitMV
 from orm_loader.helpers import Base, get_logger
 from .measurable_base import MeasurableSpec, MeasurableBase, MeasurableDomain
 
@@ -37,4 +37,15 @@ class ObserveMeasurable(DxObservationMV, MeasurableBase, Base):
         value_concept_attr="event_concept_id",
         value_string_attr="event_label",
         value_numeric_attr="value_as_number"
+    )
+
+class VisitSpecialtyMeasurable(DxRelevantVisitMV, MeasurableBase, Base):
+    __measurable__ = MeasurableSpec(
+        domain=MeasurableDomain.visit,
+        label="Diagnosis-episode linked specialty visits",
+        person_id_attr="person_id",
+        episode_id_attr="episode_id",
+        event_date_attr="visit_start_date",
+        value_concept_attr="provider_specialty_concept_id",
+        value_string_attr="provider_specialty",
     )